requirements.txt
aiogram
aiohttp
async-timeout
attrs
Babel
beautifulsoup4
bs4
certifi
chardet
idna
lxml
multidict
pytz
requests
soupsieve
typing-extensions
urllib3
yarl
pydantic
pandas
python-dotenv